src/AI func/Анализ благоустройства/requirements.txt
flask
python-dotenv
openai
pandas
PyMuPDF
openpyxl
